# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meal Market API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  school: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  school: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfile {
  id: number;
  name: string;
  school: string;
  listingCount: number;
  transactionCount: number;
  createdAt: string;
}

export type ListingStatus = (typeof ListingStatus)[keyof typeof ListingStatus];

export const ListingStatus = {
  active: "active",
  sold: "sold",
  cancelled: "cancelled",
} as const;

export interface Listing {
  id: number;
  sellerId: number;
  sellerName: string;
  sellerSchool: string;
  pointsAmount: number;
  pricePerPoint: number;
  totalPrice: number;
  /** @nullable */
  description?: string | null;
  status: ListingStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateListingBody {
  pointsAmount: number;
  pricePerPoint: number;
  description?: string;
}

export type UpdateListingBodyStatus =
  (typeof UpdateListingBodyStatus)[keyof typeof UpdateListingBodyStatus];

export const UpdateListingBodyStatus = {
  active: "active",
  sold: "sold",
  cancelled: "cancelled",
} as const;

export interface UpdateListingBody {
  pointsAmount?: number;
  pricePerPoint?: number;
  description?: string;
  status?: UpdateListingBodyStatus;
}

export interface MarketStats {
  totalActiveListings: number;
  totalPointsAvailable: number;
  avgPricePerPoint: number;
  totalTransactions: number;
  totalSchools: number;
}

export interface Conversation {
  id: number;
  listingId: number;
  listingPointsAmount: number;
  listingTotalPrice: number;
  listingStatus: string;
  buyerId: number;
  buyerName: string;
  sellerId: number;
  sellerName: string;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
  createdAt: string;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  senderName: string;
  content: string;
  /** @nullable */
  readAt?: string | null;
  createdAt: string;
}

export interface SendMessageBody {
  content: string;
}

export interface StartConversationBody {
  listingId: number;
  initialMessage: string;
}

export type TransactionRole =
  (typeof TransactionRole)[keyof typeof TransactionRole];

export const TransactionRole = {
  buyer: "buyer",
  seller: "seller",
} as const;

export interface Transaction {
  id: number;
  listingId: number;
  buyerId: number;
  buyerName: string;
  sellerId: number;
  sellerName: string;
  pointsAmount: number;
  totalPrice: number;
  role: TransactionRole;
  createdAt: string;
}

export interface CreateTransactionBody {
  listingId: number;
  buyerId: number;
}

export type GetListingsParams = {
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  sortBy?: GetListingsSortBy;
};

export type GetListingsSortBy =
  (typeof GetListingsSortBy)[keyof typeof GetListingsSortBy];

export const GetListingsSortBy = {
  newest: "newest",
  price_asc: "price_asc",
  price_desc: "price_desc",
  points_asc: "points_asc",
  points_desc: "points_desc",
} as const;