# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meal Market API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  email: zod.string(),
  password: zod.string(),
  name: zod.string(),
  school: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string(),
    school: zod.string(),
    createdAt: zod.string(),
  }),
  token: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  school: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Get all active listings
 */
export const GetListingsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  minPrice: zod.coerce.number().optional(),
  maxPrice: zod.coerce.number().optional(),
  sortBy: zod
    .enum(["newest", "price_asc", "price_desc", "points_asc", "points_desc"])
    .optional(),
});

export const GetListingsResponseItem = zod.object({
  id: zod.number(),
  sellerId: zod.number(),
  sellerName: zod.string(),
  sellerSchool: zod.string(),
  pointsAmount: zod.number(),
  pricePerPoint: zod.number(),
  totalPrice: zod.number(),
  description: zod.string().nullish(),
  status: zod.enum(["active", "sold", "cancelled"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetListingsResponse = zod.array(GetListingsResponseItem);

/**
 * @summary Create a new listing
 */
export const CreateListingBody = zod.object({
  pointsAmount: zod.number(),
  pricePerPoint: zod.number(),
  description: zod.string().optional(),
});

/**
 * @summary Get a listing by ID
 */
export const GetListingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetListingResponse = zod.object({
  id: zod.number(),
  sellerId: zod.number(),
  sellerName: zod.string(),
  sellerSchool: zod.string(),
  pointsAmount: zod.number(),
  pricePerPoint: zod.number(),
  totalPrice: zod.number(),
  description: zod.string().nullish(),
  status: zod.enum(["active", "sold", "cancelled"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a listing
 */
export const UpdateListingParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateListingBody = zod.object({
  pointsAmount: zod.number().optional(),
  pricePerPoint: zod.number().optional(),
  description: zod.string().optional(),
  status: zod.enum(["active", "sold", "cancelled"]).optional(),
});

export const UpdateListingResponse = zod.object({
  id: zod.number(),
  sellerId: zod.number(),
  sellerName: zod.string(),
  sellerSchool: zod.string(),
  pointsAmount: zod.number(),
  pricePerPoint: zod.number(),
  totalPrice: zod.number(),
  description: zod.string().nullish(),
  status: zod.enum(["active", "sold", "cancelled"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a listing
 */
export const DeleteListingParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteListingResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current user's listings
 */
export const GetMyListingsResponseItem = zod.object({
  id: zod.number(),
  sellerId: zod.number(),
  sellerName: zod.string(),
  sellerSchool: zod.string(),
  pointsAmount: zod.number(),
  pricePerPoint: zod.number(),
  totalPrice: zod.number(),
  description: zod.string().nullish(),
  status: zod.enum(["active", "sold", "cancelled"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetMyListingsResponse = zod.array(GetMyListingsResponseItem);

/**
 * @summary Get marketplace statistics
 */
export const GetListingStatsResponse = zod.object({
  totalActiveListings: zod.number(),
  totalPointsAvailable: zod.number(),
  avgPricePerPoint: zod.number(),
  totalTransactions: zod.number(),
  totalSchools: zod.number(),
});

/**
 * @summary Get all conversations for current user
 */
export const GetConversationsResponseItem = zod.object({
  id: zod.number(),
  listingId: zod.number(),
  listingPointsAmount: zod.number(),
  listingTotalPrice: zod.number(),
  listingStatus: zod.string(),
  buyerId: zod.number(),
  buyerName: zod.string(),
  sellerId: zod.number(),
  sellerName: zod.string(),
  lastMessage: zod.string().nullish(),
  lastMessageAt: zod.string().nullish(),
  unreadCount: zod.number(),
  createdAt: zod.string(),
});
export const GetConversationsResponse = zod.array(GetConversationsResponseItem);

/**
 * @summary Get messages in a conversation
 */
export const GetMessagesParams = zod.object({
  conversationId: zod.coerce.number(),
});

export const GetMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  senderId: zod.number(),
  senderName: zod.string(),
  content: zod.string(),
  readAt: zod.string().nullish(),
  createdAt: zod.string(),
});
export const GetMessagesResponse = zod.array(GetMessagesResponseItem);

/**
 * @summary Send a message in a conversation
 */
export const SendMessageParams = zod.object({
  conversationId: zod.coerce.number(),
});

export const SendMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Start a new conversation about a listing
 */
export const StartConversationBody = zod.object({
  listingId: zod.number(),
  initialMessage: zod.string(),
});

/**
 * @summary Get transaction history for current user
 */
export const GetTransactionsResponseItem = zod.object({
  id: zod.number(),
  listingId: zod.number(),
  buyerId: zod.number(),
  buyerName: zod.string(),
  sellerId: zod.number(),
  sellerName: zod.string(),
  pointsAmount: zod.number(),
  totalPrice: zod.number(),
  role: zod.enum(["buyer", "seller"]),
  createdAt: zod.string(),
});
export const GetTransactionsResponse = zod.array(GetTransactionsResponseItem);

/**
 * @summary Record a completed transaction
 */
export const CreateTransactionBody = zod.object({
  listingId: zod.number(),
  buyerId: zod.number(),
});

/**
 * @summary Get a user's public profile
 */
export const GetUserProfileParams = zod.object({
  id: zod.coerce.number(),
});

export const GetUserProfileResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  school: zod.string(),
  listingCount: zod.number(),
  transactionCount: zod.number(),
  createdAt: zod.string(),
});